TAA-1263: Android 16 upgrade
android-16.0.0_r3 + CTS updates.

+++ workloads/android/pipelines/builds/aaos_builder/aaos_environment.sh
@@ -76,7 +76,7 @@ unset BUILD_NUMBER
 # hostname: jenkins-aaos-build-pod
 
 AAOS_DEFAULT_REVISION=$(echo "${AAOS_DEFAULT_REVISION}" | xargs)
-AAOS_DEFAULT_REVISION=${AAOS_DEFAULT_REVISION:-android-16.0.0_r2}
+AAOS_DEFAULT_REVISION=${AAOS_DEFAULT_REVISION:-android-16.0.0_r3}
 
 # Android branch/tag:
 AAOS_REVISION=${AAOS_REVISION:-${AAOS_DEFAULT_REVISION}}
@@ -337,8 +337,8 @@ case "${AAOS_LUNCH_TARGET}" in
                     "curl -o .repo/local_manifests/remove_projects.xml -L ${AAOS_GERRIT_RPI_MANIFEST_URL}/android-15.0/remove_projects.xml"
                 )
                 ;;
-            *bp2a*)
-                # bp2a fallthrough: android-16.0.0_r2
+            *bp2a*|*bp3a*)
+                # bp2a/bp3a fallthrough: android-16.0.0_r2/r3
                 POST_REPO_INITIALISE_COMMANDS_LIST=(
                     "curl -o .repo/local_manifests/manifest_brcm_rpi.xml -L ${AAOS_GERRIT_RPI_MANIFEST_URL}/android-16.0/manifest_brcm_rpi.xml --create-dirs"
                     "curl -o .repo/local_manifests/remove_projects.xml -L ${AAOS_GERRIT_RPI_MANIFEST_URL}/android-16.0/remove_projects.xml"
@@ -441,12 +441,12 @@ case "${AAOS_LUNCH_TARGET}" in
                     "tail -n +315 extract-google_devices-tangorpro.sh | tar -zxvf -"
                 )
                 ;;
-            *bp2a*)
+            *bp2a*|*bp3a*)
                 echo -e "\033[1;31mTAA-1094: ${AAOS_LUNCH_TARGET} is not currently supported on ${AAOS_REVISION}!\033[0m"
                 exit 1
                 ;;
             *)
-                # android-16.0.0_r2: https://developers.google.com/android/drivers (same as bp1a above)
+                # android-16.0.0_r2/r3: https://developers.google.com/android/drivers (same as bp1a above)
                 POST_REPO_COMMAND_LIST=(
                     "curl --output - https://dl.google.com/dl/android/aosp/google_devices-tangorpro-bp1a.250505.005-fb23c626.tgz | tar -xzvf - "
                     "tail -n +315 extract-google_devices-tangorpro.sh | tar -zxvf -"
@@ -679,11 +679,13 @@ function create_workspace() {
     cd "${WORKSPACE}" || true
 
     if [[ "${ABFS_BUILDER}" == "false" ]]; then
-        # FIXME: TAA-1095 workaround - remove when fix available post android-16.0.0_r2
-        BUG_FIX="rm -rf out && ln -sf ${OUT_DIR} out"
-        echo -e "\033[1;31mTAA-1095: workaround for Android 16 OUT_DIR issue:\033[0m"
-        echo "${BUG_FIX}"
-        eval "${BUG_FIX}"
+        if [[ "${AAOS_LUNCH_TARGET}" =~ "aosp_cf_arm64" ]]; then
+            # TAA-1095 workaround
+            BUG_FIX="rm -rf out && ln -sf ${OUT_DIR} out"
+            echo -e "\033[1;31mTAA-1095: workaround for Android 16 OUT_DIR issue:\033[0m"
+            echo "${BUG_FIX}"
+            eval "${BUG_FIX}"
+        fi
     fi
 }
